Create MenuItem component and add to storybook
