+++ src/js/page_sections/Header.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import SearchSVG from 'js/svg/Search';
 import AirplaneSVG from 'js/svg/Airplane';
-import Menu from 'js/page_sections/Menu';
+import Menu from 'js/page_sections/Menu/Menu';
 import I18nLink from 'js/modules/I18nLink'
 
 class Header extends Component {
